+++ src/main/java/chess/model/board/Board.java
@@ -1,5 +1,8 @@
-package chess.model;
+package chess.model.board;
 
+import chess.model.position.File;
+import chess.model.position.Position;
+import chess.model.Team;
 import chess.model.piece.Empty;
 import chess.model.piece.Piece;
