+++ src/main/java/chess/model/piece/Rook.java
@@ -1,11 +1,9 @@
 package chess.model.piece;
 
-import chess.model.Direction;
-import chess.model.Position;
+import chess.model.position.Direction;
 import chess.model.Team;
 import chess.model.strategy.UnlimitedMoveStrategy;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class Rook extends Piece {
