+++ src/main/java/chess/model/strategy/LimitedMoveStrategy.java
@@ -1,8 +1,8 @@
 package chess.model.strategy;
 
-import chess.model.Direction;
-import chess.model.Distance;
-import chess.model.Position;
+import chess.model.position.Direction;
+import chess.model.position.Distance;
+import chess.model.position.Position;
 
 import java.util.List;
