+++ src/main/java/chess/model/strategy/MoveStrategy.java
@@ -1,6 +1,6 @@
 package chess.model.strategy;
 
-import chess.model.Position;
+import chess.model.position.Position;
 
 public interface MoveStrategy {
     boolean movable(Position source, Position target, boolean isKill);
