+++ src/test/java/chess/model/BoardTest.java
@@ -1,8 +1,11 @@
 package chess.model;
 
+import chess.model.board.Board;
+import chess.model.board.BoardFactory;
 import chess.model.piece.Pawn;
 import chess.model.piece.Piece;
 import chess.model.piece.Rook;
+import chess.model.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
