+++ src/test/java/chess/model/DistanceTest.java
@@ -1,7 +1,9 @@
 package chess.model;
 
+import chess.model.position.Direction;
+import chess.model.position.Distance;
+import chess.model.position.Position;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
