+++ src/test/java/chess/model/piece/KnightTest.java
@@ -1,6 +1,9 @@
 package chess.model.piece;
 
 import chess.model.*;
+import chess.model.board.Board;
+import chess.model.board.BoardFactory;
+import chess.model.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
